design: 디자인 수정
